+++ CHANGELOG.md
@@ -1,6 +1,11 @@
-# Change Log
+# Change Log/Notes
 
 ## Step 1
 
 - Ensure that local packages install cleanly
 - Add this changelog
+
+## Step 2
+
+- `npm install --save siwe ethers`
+- Added types/index.ts to give typing to `window.ethereum`